feat: Migrate from Jekyll to Static HTML
This commit removes the Jekyll configuration file _config.yml and replaces it with an index.html file.
This marks the first step in migrating the site from a Ruby/Jekyll setup to a purely static HTML structure. The goal is to simplify the site's architecture and remove the dependency on Ruby and Jekyll processing.

Co-authored-by: Louisy <louisy03@users.noreply.github.com>