feat: Introduce index.html for Static Site
This commit adds index.html, marking a significant step in migrating the site from a Ruby/Jekyll setup to a static HTML structure.
This file will now serve as the primary entry point for the website, containing the HTML content previously generated by Jekyll. This shift simplifies the site's architecture and eliminates the dependency on Ruby processing.

Co-authored-by: Louisy <louisy03@users.noreply.github.com>